+++ src/Microsoft.IdentityModel.Protocols/Configuration/IConfigurationEventHandler.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Microsoft.IdentityModel.Protocols.Configuration
@@ -16,15 +17,17 @@ namespace Microsoft.IdentityModel.Protocols.Configuration
         /// Called before retrieving configuration from the metadata endpoint.
         /// </summary>
         /// <param name="metadataAddress">The metadata endpoint address.</param>
+        /// <param name="cancellationToken">A cancellation token to observe while waiting for the task to complete.</param>
         /// <returns>A configuration result if available, or null to proceed with normal retrieval.</returns>
-        Task<ConfigurationEventHandlerResult<T>> BeforeRetrieveAsync(string metadataAddress);
+        Task<ConfigurationEventHandlerResult<T>> BeforeRetrieveAsync(string metadataAddress, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Called after a configuration has been successfully retrieved and updated.
         /// </summary>
         /// <param name="metadataAddress">The metadata endpoint address.</param>
         /// <param name="configuration">The updated configuration.</param>
-        Task AfterUpdateAsync(string metadataAddress, T configuration);
+        /// <param name="cancellationToken">A cancellation token to observe while waiting for the task to complete.</param>
+        Task AfterUpdateAsync(string metadataAddress, T configuration, CancellationToken cancellationToken = default);
     }
 
     /// <summary>
@@ -34,13 +37,29 @@ namespace Microsoft.IdentityModel.Protocols.Configuration
     public class ConfigurationEventHandlerResult<T> where T : class
     {
         /// <summary>
+        /// Instantiates a new instance of the <see cref="ConfigurationEventHandlerResult{T}"/> class with no result.
+        /// </summary>
+        public static readonly ConfigurationEventHandlerResult<T> NoResult = new(null, DateTimeOffset.MinValue);
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ConfigurationEventHandlerResult{T}"/> class.
+        /// </summary>
+        /// <param name="configuration">The configuration retrieved.</param>
+        /// <param name="retrievalTime"> The time when the configuration was originally retrieved (UTC).</param>
+        /// <exception cref="ArgumentNullException"></exception>
+        public ConfigurationEventHandlerResult(T configuration, DateTimeOffset retrievalTime)
+        {
+            Configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
+            RetrievalTime = retrievalTime;
+        }
+        /// <summary>
         /// Gets or sets the configuration.
         /// </summary>
-        public T Configuration { get; set; }
+        public T Configuration { get; }
 
         /// <summary>
         /// Gets or sets the time when the configuration was originally retrieved.
         /// </summary>
-        public DateTimeOffset RetrievalTime { get; set; }
+        public DateTimeOffset RetrievalTime { get; }
     }
 }
